requirements.txt
# ====================================================
# CNAE Brasil - ETL Pipeline Requirements
# ====================================================
# Projeto para processamento de dados CNAE do governo brasileiro
# Inclui download, processamento e análise de empresas, estabelecimentos e sócios

# Core Data Processing
pandas>=2.2.0              # Manipulação de dados e DataFrames
duckdb>=0.10.0             # Análise SQL performante e conversão Parquet  
pyarrow>=15.0.0            # Suporte Parquet e Arrow (melhor integração)

# HTTP Requests & Downloads
requests>=2.31.0           # Download de arquivos do governo
urllib3>=2.0.0             # Dependência do requests (segurança)

# Progress Bars & User Experience  
tqdm>=4.66.0               # Barras de progresso durante processamento

# Date/Time Handling
python-dateutil>=2.8.2     # Cálculos de datas para URLs dinâmicas

# Encoding Detection
chardet>=5.2.0             # Detecção automática de encoding (CSV brasileiro)

# Development & Debugging (opcional)
# ipython>=8.0.0           # Shell interativo melhorado
# jupyter>=1.0.0           # Notebooks para análise exploratória